Adding two new files
adding employee login and password reset, employee login is a page for employees to log in, maybe it should be combined with customer login, but right now they are very similar to each other besides the table they reference. password reset should be usable for both employee's and customers as it checks if user id or customer id is set before changing the password of the associated account.